+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+2.0.31 (2017-03-22)
+-------------------
+
+    * Assignment: PM and Scouters can schedule and re-schedule assignments in the past (ericof).
+
+
 2.0.30 (2017-03-19)
 -------------------